pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "csfpy"
version = "0.1.0"
description = "An Analytical Engine for Non-Prismatic Structural Members"
readme = "README.md"
authors = [{ name = "Giovanni Boscu" }]
license = { text = "GPL-3.0" }
requires-python = ">=3.8"
keywords = ["structural-mechanics", "analytical-engine", "non-prismatic-beams", "civil-engineering"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy",
    "matplotlib",
    "PyYAML>=6.0",
    "opensees; sys_platform == 'win32'",
    "openseespy; sys_platform != 'win32'"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["csf*"]

[project.optional-dependencies]
examples = [
    "pycba"
]
